pyside: Include C development headers and config
The Python installer doesn't include the necessary headers for C
development. Manually include the necessary headers and configuration
files. Other distributions such as Ubuntu already do this when they
build the -dev packages.

This is a slight change of direction from the original 5.12.2 package
(#38180) which tries to be as close as possible to the version in
pip. However, there is software out there that depends on the C
interfaces. For example, FreeCAD's official tap includes a "private"
PySide version that includes these components.

Signed-off-by: Andreas Sandberg <andreas@sandberg.uk>